+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 ## [Unreleased]
 
+### Fixed
+
+* [MIC-6177] - Card field's input mask for PAN's with dynamic length
+
 ## [2.9.0] - 2022-08-12Z
 
 ### Fixed